requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
